fix: serializable parameters
